+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # main
 - Bugfix:           ConditionFileExist: fixed bug with wildcards.
+- Bugfix:           GUI: fixed MatchCaseSensitive for schema selction.
 - Bugfix:           GUI: fixed bug when moving elements to same position.
 - Bugfix:           GNSS, WriteNormalEquations: check non-matching parameter selection.
 - Other:            NetCDF: enable to read multidimensional variables to store in muliple columns.